requirements.txt
pypdf
langchain
langchain-community
langchain-ollama
langchain-chroma
chromadb
pytest
boto3
langdetect
sentence-transformers
transformers
torch
whisper
pyaudio
indic-nlp-library
huggingface-hub
deep-translator
psutil 